requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
alembic>=1.14.0
psycopg2-binary>=2.9.0
pydantic-settings>=2.6.0
stripe>=11.0.0
sib-api-v3-sdk>=7.6.0
slowapi>=0.1.9
kerykeion>=5.7.0
pyswisseph>=2.10.3.2
reportlab>=4.2.0
geopy>=2.4.1
timezonefinder>=6.5.0
PyYAML>=6.0.2
python-dateutil>=2.9.0
httpx>=0.27.0
pytest>=8.3.0
